movie_review_app/pyproject.toml
[tool.poetry]
name = "movie-review-app"
version = "0.1.0"
description = "Movie Review FastAPI App"
authors = ["You <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = "*"
tortoise-orm = "*"
aerich = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
